+++ oap-server/server-query-plugin/logql-plugin/src/main/java/org/apache/skywalking/oap/query/logql/LogQLProvider.java
@@ -73,6 +73,9 @@ public class LogQLProvider extends ModuleProvider {
                                                             .contextPath(config.getRestContextPath())
                                                             .idleTimeOut(config.getRestIdleTimeOut())
                                                             .acceptQueueSize(config.getRestAcceptQueueSize())
+                                                            .enableTLS(config.isRestSSLEnabled())
+                                                            .tlsKeyPath(config.getRestSSLKeyPath())
+                                                            .tlsCertChainPath(config.getRestSSLCertChainPath())
                                                             .build();
 
         httpServer = new HTTPServer(httpServerConfig);