+++ oap-server/server-query-plugin/traceql-plugin/src/main/java/org/apache/skywalking/oap/query/traceql/TraceQLProvider.java
@@ -75,6 +75,9 @@ public class TraceQLProvider extends ModuleProvider {
                                                                 "/") // Base context path for the server, individual handlers will have their own context paths
                                                             .idleTimeOut(config.getRestIdleTimeOut())
                                                             .acceptQueueSize(config.getRestAcceptQueueSize())
+                                                            .enableTLS(config.isRestSSLEnabled())
+                                                            .tlsKeyPath(config.getRestSSLKeyPath())
+                                                            .tlsCertChainPath(config.getRestSSLCertChainPath())
                                                             .build();
 
         httpServer = new HTTPServer(httpServerConfig);